Created navigation component to reduce code duplication and q&a fixes
